[IMP] accounting: default_taxes.rst - Changes on invoices
Since https://github.com/odoo/odoo/pull/115222, a new line on an invoice won't have
the default sales tax put on by default. If there are no taxes on
the product, the account or from predictive, the tax will stay empty
Its use has been restricted back to the tax put on new products.
It is still used in Accounting Firms mode, as we don't want people
to have to input the product when quick encoding.

task-3045026